# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Articles API",
        "contact": {
            "name": "Support Team",
            "email": "don.unobunjo@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/articles": {
            "get": {
                "operationId": "/articles",
                "parameters": [
                    {
                        "name": "articles",
                        "in": "path",
                        "description": "The articles parameter in path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fetch all articles",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
        "/articles/{id}": {
            "get": {
                "tags": [
                    "yourtag"
                ],
                "operationId": "/articles/{id}",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns an article",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        }
    }
}